Testing Done & Ready To Submit

+++ frontend/src/app/pages/user/user-address/user-address.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { LoginService } from 'src/app/services/login.service';
 import { UserService } from 'src/app/services/user.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-user-address',
@@ -28,15 +29,33 @@ export class UserAddressComponent implements OnInit {
   }
 
   deleteAddress(address_id:any){
+
+Swal.fire({
+  title: 'Are you sure?',
+  text: 'You Want To Delete Address!',
+  icon: 'question',
+  showCancelButton: true,
+  confirmButtonColor: '#3085d6',
+  cancelButtonColor: '#d33',
+  confirmButtonText: 'Yes, delete it!',
+}).then((result) => {
+  if (result.isConfirmed) {
     this.userService.deleteAddress(address_id).subscribe(
-      (data)=>{
-        this.snack.open('Address deleted successfully','OK');
+      (data) => {
         location.reload();
+        Swal.fire('Address deleted successfully!', '', 'success');
+        this.userService.cartStatus.next(true);
       },
-      (error)=>{
-        this.snack.open('Address not deleted', 'OK');
+      (error) => {
+        this.snack.open("You can't delete this address rigth now.Already order is pending on same address.", 'OK');
       }
-    )
+    );
+      
+ 
+  }
+});
+
+    
   }
 
 }
